+++ x/leverage/keeper/genesis.go
@@ -11,7 +11,9 @@ func (k Keeper) InitGenesis(ctx sdk.Context, genState types.GenesisState) {
 	k.SetParams(ctx, genState.Params)
 
 	for _, token := range genState.Registry {
-		k.SetRegisteredToken(ctx, token)
+		if err := k.SetRegisteredToken(ctx, token); err != nil {
+			panic(err)
+		}
 	}
 
 	for _, borrow := range genState.AdjustedBorrows {