+++ x/leverage/keeper/iter.go
@@ -196,8 +196,8 @@ func (k Keeper) GetEligibleLiquidationTargets(ctx sdk.Context) ([]sdk.AccAddress
 			return err
 		}
 
-		// compute liquidation limit from enabled collateral
-		liquidationLimit, err := k.CalculateLiquidationLimit(ctx, collateral)
+		// compute liquidation threshold from enabled collateral
+		liquidationLimit, err := k.CalculateLiquidationThreshold(ctx, collateral)
 		if err != nil {
 			return err
 		}